User cann't update or delete another article

+++ article-posting/src/routers/article.js
@@ -5,25 +5,32 @@ const auth = require('../middleware/auth')
 
 const router = new express.Router()
 
-router.post("/articles/new", async (req, res) => {
-    const article = new Article(req.body)
+router.post("/articles/new", auth, async (req, res) => {
+    const article = new Article(
+        {
+            ...req.body,
+            author: req.user._id
+        }
+    )
 
     try {
         await article.save()
+        await req.user.articles.unshift(article)
+        await req.user.save()
         res.status(201).send({ article })
     } catch (e) {
-        res.status(400).send(e.message)
+        res.status(400).send({ error: 'From Router : ' + e.message })
     }
 })
 
 router.get("/articles/all", async (req, res) => {
-    Article.find({}).populate('comments').exec((err, result) => {
+    Article.find({}).populate('comments').populate('author').exec((err, result) => {
         res.send(result)
     })
 })
 
 router.get("/articles/:id", async (req, res) => {
-    Article.findById(req.params.id).populate('comments').exec((err, result) => {
+    Article.find({ author: req.params.id }).populate('comments').exec((err, result) => {
         if (err) {
             res.send(err)
         }
@@ -32,9 +39,9 @@ router.get("/articles/:id", async (req, res) => {
     })
 })
 
-router.patch('/articles/:id', async (req, res) => {
+router.patch('/articles/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body)
-    const allowedUpdates = ["title", "content"]
+    const allowedUpdates = ["title", "topic", "content"]
     const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
 
     if (!isValidOperation) {
@@ -42,7 +49,7 @@ router.patch('/articles/:id', async (req, res) => {
     }
 
     try {
-        const result = await Article.findOne({ _id: req.params.id })
+        const result = await Article.findOne({ _id: req.params.id, author: req.user._id })
 
         updates.forEach((update) => result[update] = req.body[update])
 
@@ -54,15 +61,16 @@ router.patch('/articles/:id', async (req, res) => {
         }
 
         res.send(result)
+        
     } catch (e) {
         console.log(e);
         res.status(400).send()
     }
 })
 
-router.delete('/articles/:id', async (req, res) => {
+router.delete('/articles/:id', auth, async (req, res) => {
     try {
-        const article = await Article.findOneAndDelete({ _id: req.params.id })
+        const article = await Article.findOneAndDelete({ _id: req.params.id, author: req.user._id })
 
         if (!article) {
             return res.status(404).send()
